+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - The `create` subcommand now supports a `--small`/`-s` flag that reduces the output down to only core components and common fields.
+- Go API: Added method `Overlay` to the public Bloblang package.
 
 ## 3.45.1 - 2021-04-27